move new item types to shared helper file
